refactor: refactor ItinerariesList component
Extract list into seperate component and refactor getItineraries function to accept an optional user id
