+++ src/app/lib/data/itineraries.js
@@ -79,22 +79,22 @@ export async function getItineraryById(id) {
   }
 }
 
-export async function getItineraries(limit) {
+export async function getItineraries(limit,user_id) {
   try {
+    let params = [];
     let sqlString = `
-    SELECT i.itinerary_id, i.title, i.itinerary_image_url, i.user_id, i.itinerary_description, i.created_at, i.budget, COALESCE(CAST(COUNT(d.day_number)AS INTEGER),0) AS number_of_days, uv.username, uv.total_votes, loc.country_list, loc.region_list, loc.place_list
-    FROM itineraries i 
+    SELECT i.itinerary_id, i.title, i.itinerary_image_url, i.user_id, i.itinerary_description, i.created_at, i.budget, COALESCE(CAST(COUNT(d.day_number)AS INTEGER),0) AS number_of_days, u.username, uv.total_votes, loc.country_list, loc.region_list, loc.place_list
+    FROM itineraries i
     LEFT JOIN days d
     ON i.itinerary_id=d.itinerary_id
-
+    LEFT JOIN users u
+    ON i.user_id=u.user_id
     LEFT JOIN (
-      SELECT i.itinerary_id, u.username, COALESCE(CAST(SUM(vote_value)AS INTEGER),0) AS total_votes
+      SELECT i.itinerary_id, COALESCE(CAST(SUM(vote_value)AS INTEGER),0) AS total_votes
       FROM itineraries i
       FULL JOIN itinerary_votes v
       ON i.itinerary_id=v.itinerary_id
-      INNER JOIN users u
-      ON i.user_id=u.user_id
-      GROUP BY i.itinerary_id, u.username
+      GROUP BY i.itinerary_id
     ) uv
     ON i.itinerary_id=uv.itinerary_id
 
@@ -106,14 +106,18 @@ export async function getItineraries(limit) {
       GROUP BY i.itinerary_id
     ) loc
     ON i.itinerary_id=loc.itinerary_id
-    GROUP BY i.itinerary_id, uv.username, uv.total_votes, loc.country_list, loc.region_list, loc.place_list
     `;
-
-    if (limit) {
-      sqlString += `LIMIT ${limit}`;
+    if (user_id) {
+      params.push(user_id)
+      sqlString += `WHERE i.user_id = $${params.length} `
     }
-    const allItinerariesSQL = await sql.query(sqlString);
-
+sqlString += `GROUP BY i.itinerary_id, u.username, uv.total_votes, loc.country_list, loc.region_list, loc.place_list
+    ORDER BY total_votes DESC `
+    if (limit) {
+      params.push(limit)
+      sqlString += `LIMIT $${params.length}`;
+    }
+    const allItinerariesSQL = await sql.query(sqlString,params);
 
     return allItinerariesSQL.rows;
   } catch (error) {
